+++ src/brr.mli
@@ -3340,15 +3340,23 @@ module Window : sig
 
   val inner_width : t -> int
   (** [inner_width w] is the
-      {{:https://developer.mozilla.org/en-US/docs/Web/API/Window/innerWidth}interior
-      height} of the window in CSS pixels, including the width of the vertical
-      scroll bar, if present. *)
+      {{:https://developer.mozilla.org/en-US/docs/Web/API/Window/innerWidth}
+      interior height} of the window in CSS pixels, including the width of
+      the vertical scroll bar, if present. *)
 
   val inner_height : t -> int
   (** [inner_height w] is the
-      {{:https://developer.mozilla.org/en-US/docs/Web/API/Window/innerHeight}interior
-      height} of the window in CSS pixels, including the height of the horizontal
-      scroll bar, if present. *)
+      {{:https://developer.mozilla.org/en-US/docs/Web/API/Window/innerHeight}
+      interior height} of the window in CSS pixels, including the height of
+      the horizontal scroll bar, if present. *)
+
+  val parent : t -> t option
+  (** [parent w] is the
+      {{:https://developer.mozilla.org/en-US/docs/Web/API/Window/parent}parent}
+      of the window, if it has one.
+
+      When a window is loaded in an [<iframe>], [<object>], or [<frame>], its
+      parent is the window with the element embedding the window. *)
 
   (** {1:media Media properties} *)